+++ subsys/sd/CMakeLists.txt
@@ -7,4 +7,6 @@ if (CONFIG_SD_STACK)
   zephyr_library_sources(sd.c sd_ops.c)
   zephyr_library_sources_ifdef(CONFIG_SDMMC_STACK sdmmc.c)
   zephyr_library_sources_ifdef(CONFIG_SDIO_STACK sdio.c)
+  zephyr_library_sources_ifdef(CONFIG_MMC_STACK mmc.c)
+
 endif()
